fix(models): break row projection import cycle
